Move I2C class to HAL
