+++ inc/hal/MicroBitI2C.h
@@ -14,22 +14,22 @@
 class MicroBitI2C : public I2C
 {
     uint8_t retries;
-    
+
     public:
-    
+
     /**
-      * Constructor. 
+      * Constructor.
       * Create an instance of i2c
       * @param sda the Pin to be used for SDA
       * @param scl the Pin to be used for SCL
       * Example:
-      * @code 
+      * @code
       * MicroBitI2C i2c(MICROBIT_PIN_SDA, MICROBIT_PIN_SCL);
       * @endcode
       * @note this should prevent i2c lockups as well.
       */
     MicroBitI2C(PinName sda, PinName scl);
-  
+
     /**
       * Performs a complete read transaction. The bottom bit of the address is forced to 1 to indicate a read.
       *
@@ -41,12 +41,12 @@ class MicroBitI2C : public I2C
       * @return MICROBIT_OK on success, MICROBIT_I2C_ERROR if an unresolved read failure is detected.
       */
     int read(int address, char *data, int length, bool repeated = false);
-   
+
     /**
       * Performs a complete write transaction. The bottom bit of the address is forced to 0 to indicate a write.
       *
       * @address 8-bit I2C slave address [ addr | 0 ]
-      * @data Pointer to the byte-arraycontaining the data to write 
+      * @data Pointer to the byte-arraycontaining the data to write
       * @length Number of bytes to write
       * @repeated Repeated start, true - don't send stop at end.
       *
