+++ source/hal/MicroBitHal.cpp
@@ -1,0 +1,5 @@
+
+#include "hal/MicroBitHal.h"
+
+MicroBitI2C MicroBitHal::i2c(MICROBIT_PIN_SDA, MICROBIT_PIN_SCL);
+
